tokenize post comments

+++ api/v1/routes/post_comments.py
@@ -2,6 +2,7 @@ from flask import Blueprint, jsonify, request, make_response, abort
 from models.post import Post, PostComment
 from models import storage
 from flasgger.utils import swag_from
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 comments_bp = Blueprint('postComments', __name__)
 
@@ -9,20 +10,32 @@ comments_bp = Blueprint('postComments', __name__)
 # Retrieves all comments for a specific post
 @swag_from('documentation/post/comments/get_post_comments.yml')
 @comments_bp.route('/posts/<post_id>/comments', methods=['GET'])
+@jwt_required()
 def get_post_comments(post_id):
-    post =  storage.get(Post, post_id)
+    """retrieves a list of all comments for a specific post """
+    post = storage.get(Post, post_id)
     return jsonify([comment.to_dict() for comment in post.comments])
 
 
 # Creates a new comment for a specific post
 @swag_from('documentation/post/comments/create_post_comment.yml')
 @comments_bp.route('/posts/<post_id>/comments', methods=['POST'])
+@jwt_required()
 def create_post_comment(post_id):
+    """creates a new comment for a specific post """
     if not request.get_json():
         abort(400, description="Not a JSON")
 
     if 'text' not in request.get_json():
         abort(400, description="Missing text")
+
+    if 'user_id' not in request.get_json():
+        abort(400, description="Missing user_id")
+
+    current_user_id = get_jwt_identity()
+
+    if request.get_json()['user_id'] != current_user_id:
+        abort(403, "Permission denied, not the current logged in user")
 
     data = request.get_json()
     instance = PostComment(**data)
@@ -34,11 +47,18 @@ def create_post_comment(post_id):
 # Updates a comment for a specific post
 @swag_from('documentation/post/comments/update_post_comment.yml')
 @comments_bp.route('/posts/<post_id>/comments/<comment_id>', methods=['PUT'])
+@jwt_required()
 def update_post_comment(post_id, comment_id):
-    postComment =  storage.get(PostComment, comment_id)
+    """update a comment for a specific post"""
+    postComment = storage.get(PostComment, comment_id)
 
     if not request.get_json():
         abort(400, description="Not a JSON")
+
+    current_user_id = get_jwt_identity()
+
+    if postComment.user_id != current_user_id:
+        abort(403, "Permission denied, not the current logged in user")
 
     ignore = ['id', 'user_id', 'created_at', 'updated_at']
 
@@ -53,8 +73,15 @@ def update_post_comment(post_id, comment_id):
 # Deletes a comment for a specific post
 @swag_from('documentation/post/comments/delete_post_comment.yml')
 @comments_bp.route('/posts/<post_id>/comments/<comment_id>', methods=['DELETE'])
+@jwt_required()
 def delete_post_comment(post_id, comment_id):
-    postComment =  storage.get(PostComment, comment_id)
+    """deletes a comment for a specific post"""
+    current_user_id = get_jwt_identity()
+
+    postComment = storage.get(PostComment, comment_id)
+    if postComment.user_id != current_user_id:
+        abort(403, "Permission denied, not the current logged in user")
+
     storage.delete(postComment)
     storage.save()
     return make_response(jsonify({"message": "Comment deleted"}), 200)
